jurisprudencia_scraper/requirements.txt
selenium==4.26.1
undetected-chromedriver==3.5.5
beautifulsoup4==4.12.3
lxml==5.3.0
psycopg2-binary==2.9.10
redis==5.2.0
SQLAlchemy==2.0.36
python-dotenv==1.0.1
requests==2.32.3
pandas==2.2.3
loguru==0.7.2
tenacity==9.0.0
colorama==0.4.6
tqdm==4.67.0
numpy==2.1.3